Render parsermd objects

Object contents are converted to a character vector and written to a temporary directory before rendering via quarto::quarto_render() or rmarkdown::render().

Note that this function has the potential to overwrite existing output files (e.g. .html, .pdf, etc).

Usage

render(x, name = NULL, ..., engine = c("quarto", "rmarkdown"))

Arguments

x

Object to render, e.g. a rmd_ast, rmd_tibble, character vector, etc.

name

Name of the output file, if not given it will be inferred from the name of x.

...

Any additional arguments to be passed to quarto::quarto_render() or rmarkdown::render()

engine

The rendering engine to use, either "quarto" or "rmarkdown".

Value

Returns the results of the render function.