Rmd selection helper functions

These functions are used in conjunction with rmd_select() to select nodes from an Rmd ast.

has_type() - selects all nodes that have the given type(s).

by_section() - uses section selectors to select nodes.

has_type(types)

by_section(sec_ref, keep_parents = TRUE)

by_label(label)

has_option(...)

Arguments

types
Vector of character type names, e.g. rmd_chunk, rmd_heading, etc.

sec_ref
character vector, a section reference selector. See details below for further details on how these are constructed.

keep_parents
Logical, retain the parent headings of selected sections. Default: TRUE

label
character vector, glob patterns for matching chunk labels.

...
Either option names represented by a scalar string or a named argument with the form opt = value where opt is the option name and value is the value to be checked. For example eval = TRUE would check for the option eval being set to TRUE.

Details

Section reference selectors

Section reference selectors are a simplified version of CSS selectors that are designed to enable the selection nodes in a way that respects the implied hierarchy of a document's section headings.

They consist of a character vector of heading names where each subsequent value is assumed to be nested within the preceding value. For example, the section selector c("Sec 1", "Sec 2") would select all nodes that are contained within a section named Sec 2 that is in turn contained within a section named Sec 1 (or a section contained within a section named Sec 1, and so on).

The individual section names can be specified using wildcards (aka globbing patterns), which may match one or more sections within the document, e.g. c("Sec 1", "Sec *"). See utils::glob2rx() or wikipedia for more details on the syntax for these patterns.